+++ aim/web/ui/src/pages/RunDetail/DistributionsVisualizer/DistributionVisualizer.tsx
@@ -2,7 +2,6 @@ import React, { memo } from 'react';
 
 import { IDistributionVisualizerProps } from '../types';
 
-import RangePanel from './RangePanel';
 import Wrapper from './temp/Wrapper';
 
 import './DistributionsVisualizer.scss';
@@ -13,23 +12,7 @@ function DistributionsVisualizer(
   return (
     <div className='DistributionsVisualizer'>
       {props.data?.processedValues && (
-        <>
-          <Wrapper data={props.data?.processedValues} className='Visualizer' />
-          <RangePanel
-            items={[
-              {
-                name: 'density',
-                sliderTitle: 'Test1',
-                inputTitle: 'Test1',
-                sliderTitleTooltip: 'test1-tooltip',
-                inputTitleTooltip: 'test1-tooltip',
-                rangeEndpoints: [10, 20],
-                selectedRangeValue: [10, 20],
-                inputValue: 10,
-              },
-            ]}
-          />
-        </>
+        <Wrapper data={props.data?.processedValues} className='Visualizer' />
       )}
     </div>
   );