+++ aim/web/ui/src/pages/RunDetail/TraceVisualizationContainer/TraceVisualizationContainer.tsx
@@ -10,6 +10,7 @@ import runTracesModel from 'services/models/runs/runTracesModel';
 import DistributionsVisualizer from '../DistributionsVisualizer';
 import ImagesVisualizer from '../ImagesVisualizer/ImagesVisualizer';
 import { ITraceVisualizationContainerProps } from '../types';
+import RangePanel from '../RangePanel';
 
 import widthEmptyTraceCheck from './widthEmptyTraceCheck';
 
@@ -18,6 +19,10 @@ import './TraceVisualizationContainer.scss';
 const traceTypeVisualization = {
   images: ImagesVisualizer,
   distributions: DistributionsVisualizer,
+  audios: () => null,
+  videos: () => null,
+  texts: () => null,
+  plots: () => null,
 };
 
 function TraceVisualizationContainer({
@@ -60,6 +65,31 @@ function TraceVisualizationContainer({
             isLoading={runTracesModelData?.isTraceBatchLoading}
             activeTraceContext={runTracesModelData?.menu?.activeItemName}
           />
+
+          {runTracesModelData?.data &&
+            runTracesModelData?.config &&
+            runTracesModelData?.queryData && (
+              <RangePanel
+                items={runTracesModelData?.config?.rangePanel.map((item) => ({
+                  key: item.sliderName,
+                  sliderName: item.sliderName,
+                  inputName: item.inputName,
+                  sliderTitle: item.sliderTitle,
+                  inputTitle: item.inputTitle,
+                  sliderTitleTooltip: item.sliderTitleTooltip,
+                  inputTitleTooltip: item.inputTitleTooltip,
+                  rangeEndpoints: runTracesModelData?.data[item.sliderName],
+                  selectedRangeValue: runTracesModelData?.queryData?.sliders[
+                    item.sliderName
+                  ] || [0, 50],
+                  inputValue:
+                    runTracesModelData?.queryData?.inputs[item.inputName] || 50,
+                }))}
+                onApply={runTracesModel.onApply}
+                onInputChange={runTracesModel.onInputChange}
+                onRangeSliderChange={runTracesModel.onRangeChange}
+              />
+            )}
         </BusyLoaderWrapper>
       </div>
     </div>